+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.13.19] - (25.04.2022)
+## Changes
+- refactor send debug and info log
+- refactor send mqtt message
+- update dependencies
+
 ## [3.13.16] - (24.04.2022)
 ## Changes
 - fixed MQTT info report